demo

+++ app/controllers/userController.js
@@ -67,6 +67,11 @@ exports.deleteUser = async (req, res) => {
     if (!user) {
         return res.status(400).json({ 'message': `User ID ${req.body.id} not found` });
     }
-    const result = await user.deleteOne({ _id: req.body.id });
-    res.json(result);
+    
+    if (user.roles === 0) {
+        const result = await user.deleteOne({ _id: req.body.id });
+        res.json(result);
+    } else {
+        return res.status(400).json({ 'message': 'Không có quyền xóa user này!' });
+    }
 }
